+++ resources/views/components/layouts/app.blade.php
@@ -553,6 +553,7 @@
     </script>
     @stack('scripts')
     <x-billing::paywall-overlay />
+    @include('notification::partials.popup-modal')
 </body>
 
 </html>